refactor: Separate tab rendering from logic
